Fix bug when changing value of correct answer
When changing the value of the correct answer of a question with
multiple possible answers, the action would take a long time to
complete.

This happened because multiple 'change' event listeners were added to
the drop down list in a loop. Only one event listener to the container
div needs to be added, and the container's children will automatically
be registered for the 'change' event.