Make toc a plugin again
The toc needs to be a plugin in order for it to function outside of the
Butter environment (for example, the exported page).

The plugin can only be edited from clicking on the "Open TOC Editor"
button; the track event for the toc is hidden to enforce this.

The toc plugin is also hidden from the plugin-list editor in order to
ensure that only one toc is added to the project.