Improve default state of TOC
The default now includes common sections used by Fivel:

1) Introduction
2) Insight 1
3) Insight 2
4) Insight 3
5) Insight 4
6) Insight 5
7) Review
8) Quiz

The time for a section can now be specified as {{duration}}, which
specifies to the TOC that the section should use the current video's
duration.

The sorting now only happens when the TOC editor opens, which makes
editing a section easier.

Fixes #60.